backend/requirements.txt
# FastAPI 框架
fastapi==0.109.0
uvicorn[standard]==0.27.0

# 数据库
sqlalchemy==2.0.25
aiosqlite==0.19.0

# 爬虫
requests==2.31.0
beautifulsoup4==4.12.3
lxml==5.1.0

# 美股数据
yfinance==0.2.36

# 定时任务
apscheduler==3.10.4

# 工具
python-dotenv==1.0.0
pydantic==2.5.3
pydantic-settings==2.1.0
httpx==0.26.0

# 日期处理
python-dateutil==2.8.2
pytz==2024.1
